unit-test: updating the Hop CI tests
