+++ tests/kLASSCF/test_pbc_klasscf_hessian_ci.py
@@ -1,16 +1,16 @@
-"""Tests for k-LASSCF CI updates, Hessian responses, and vector layouts.
-
-Eight focused tests cover combined orbital/CI updates, complex CI tangent
-rotations, diagonal and off-diagonal CI Hessian actions, Hessian-diagonal
-assembly, and complex gradient construction.
-"""
-
 import unittest
 import numpy as np
 
 from mrh.my_pyscf.pbc.mcscf.klasscf import KLASSCF_HessianOperator
 
 # Author: Bhavnesh Jangid
+
+"""Tests for k-LASSCF CI updates, Hessian responses, and vector layouts.
+
+Eleven focused tests cover packed complex CI layouts, combined orbital/CI
+updates, complex CI tangent rotations, diagonal and off-diagonal CI Hessian
+actions, Hessian-diagonal assembly, and complex gradient construction.
+"""
 
 class _RecordingFCIBox:
 
@@ -64,6 +64,39 @@ class _RecordingStepUGG:
         return self.kappa, self.dci
 
 
+class _IdentityTransformer:
+    ncsf = 2
+    ndet = 2
+
+    @staticmethod
+    def vec_det2csf(vector, order="C", normalize=False):
+        return np.array(vector, copy=True)
+
+    @staticmethod
+    def vec_csf2det(vector, order="C", normalize=False):
+        return np.array(vector, copy=True)
+
+
+class _FakeUGG:
+    nvar_tot = 7
+
+
+def make_layout_operator(frozen_ci=()):
+    operator = KLASSCF_HessianOperator.__new__(KLASSCF_HessianOperator)
+    operator.ugg = _FakeUGG()
+    operator.ci = [
+        [np.array([1.0, 0.0], dtype=np.complex128)],
+        [np.array([0.0, 1.0], dtype=np.complex128)],
+    ]
+    operator.ci_transformers = [
+        [_IdentityTransformer()],
+        [_IdentityTransformer()],
+    ]
+    operator.frozen_ci = set(frozen_ci)
+    operator.nvar_ci = 2 * (2 - len(operator.frozen_ci))
+    return operator
+
+
 def make_diagonal_operator():
     operator = KLASSCF_HessianOperator.__new__(KLASSCF_HessianOperator)
     boxes = [
@@ -94,6 +127,44 @@ def make_diagonal_operator():
 
 
 class KnownValues(unittest.TestCase):
+
+    def test_complex_csf_layout_round_trip(self):
+        """Unpack complex CSF coordinates into determinant CI vectors.
+
+        Repacking must reproduce the original coordinates, and the unpacked
+        vectors must retain the expected fragment/root layout.
+        """
+        operator = make_layout_operator()
+        packed = np.array([
+            0.2 + 0.1j, -0.3j, 0.4 - 0.2j, 0.5j,
+        ])
+
+        ci = operator._unpack_ci_vector(packed)
+
+        np.testing.assert_allclose(operator._flatten_ci_vector(ci), packed)
+        self.assertEqual(len(ci), 2)
+        self.assertEqual(len(ci[0]), 1)
+
+    def test_frozen_ci_is_zero_and_omitted(self):
+        """Omit frozen CI variables when packing and restore zero responses."""
+        operator = make_layout_operator(frozen_ci=(1,))
+        packed = np.array([0.2 + 0.1j, -0.3j])
+
+        ci = operator._unpack_ci_vector(packed)
+
+        np.testing.assert_allclose(ci[0][0], packed)
+        np.testing.assert_array_equal(ci[1][0], np.zeros(2))
+        np.testing.assert_allclose(operator._flatten_ci_vector(ci), packed)
+
+    def test_zero_ci_helpers_preserve_layout_and_dtype(self):
+        """Create and detect zero CI steps without changing layout or dtype."""
+        operator = make_layout_operator()
+        zero = operator._zero_ci_step(np.complex128)
+
+        self.assertTrue(operator._ci_step_is_zero(zero))
+        self.assertTrue(np.issubdtype(zero[0][0].dtype, np.complexfloating))
+        zero[1][0][0] = 1.0j
+        self.assertFalse(operator._ci_step_is_zero(zero))
 
     def test_combined_update_dispatches_packed_orbital_and_ci_rotations(self):
         """Dispatch unpacked orbital and CI steps to their update methods."""
